./tool test --max-errored-tests
Summary:
This adds two flags to `./tool test`. The idea is to set a
maximum number of test suites which are allowed to error (not fail). If
no tests fail and we don't exceed the maximum number of errored tests,
then we can exit with exit code 0.

You can set an absolute maximum with `--max-errored-tests` and a
percentage maximum with `--max-errored-tests-pct`.

This should make AppVeyor more reliable until we make Flow on Windows
more reliable. We can set the threshold to something like 20% (often
like 3 out of 33 tests error).

Reviewed By: samwgoldman

Differential Revision: D7415246

fbshipit-source-id: d5d656bdbd17d7dc4edd1d1ae250542cefac1998