Recover from printing errors when contents have changed
Summary:
Flow errors reference code in files. We read those files when
pretty printing the errors in order to show context. If those files have
changed since the error was generated, the pretty printing might throw
an exception.

This diff tries to avoid throwing those exceptions by failing to render
the code frame if the line lengths are too short.

It also fixes a second bug when printing colorless code frames. When
printing multiline locations, we underline the end of the code block.
This could was triggering exceptions too.

Reviewed By: samwgoldman, calebmer

Differential Revision: D7415778

fbshipit-source-id: fc2c8814f28c071afb86a248ecc0fda458b00853