ObjT ~> ObjT hook
Summary:
This will be used, to begin with, to link object literals with types that they flow into for use in find-refs. Eventually it will be used to link arbitrary object types together for a more expansive version of find-refs.

I'd like to get rid of most of the hooks, and replace them with calls into the type table, but I don't believe I can do that here. The type table does not store relationships between types. Therefore we need this hook to accomplish our find-refs goals.

Maybe in the future, the type table could store subtype relationships for types, at which point we could remove this hook. That would likely make the type table blow up in size though, so for now this is the efficient and expedient way to go.

Reviewed By: panagosg7

Differential Revision: D7386252

fbshipit-source-id: 1ff3fbc010f645218bdcda9c31ae3fed26ccc84e