Collect object dictionaries for full type resolution
Summary:
The type of an indexer might influence which arm we choose during speculation.
Before this change, we did not collect tvars associated with indexers. Since we
did not collect these tvars, we did not defer the related actions, which caused
spurious errors.

In the added test, before this diff Flow would error that `string p` was
incompatible with the `number` indexer in the argument. After this diff, that
constraint is deferred, so the error becomes a speculation failure and we
choose the second arm.

Reviewed By: panagosg7

Differential Revision: D7404688

fbshipit-source-id: 733c0a2ea6175811bbb11c366df0125dc42f92b7