[compiler] avoid unnecessary Sequences
Summary:
`list` wraps everything in an additional `Sequence` which isn't necessary in some cases and `Break_if_needed` causes expensive backtracking. The right long term fix is to make `Break_if_needed` not expensive, but we can also avoid doing extra work.

the extra `Sequence` is to handle cases where `wrap` contains an `IfBreak` so we need to reset the break mode to prevent it from adding the wrappers unnecessarily. but if we know `wrap` doesn't contain `IfBreak`, then we don't need the extra `Sequence`.

Reviewed By: pieterv

Differential Revision: D7445421

fbshipit-source-id: d96592e64c1a1194f0e9b5d006c5e068c7e15e9b