Add locs to exports
Summary:
Missing location information for properties was causing `find-refs` to fail in certain circumstances. This adds additional location information and a test to exercise it.

Unfortunately, this still doesn't work well with global find-refs because of the split between variable find-refs and object type find-refs. This should get a bit easier to sort out once I unify the File_sig import/export calculations with those of the type system.

Reviewed By: samwgoldman

Differential Revision: D7451879

fbshipit-source-id: f42c96782016cdea1bd99bfdcb36b45d1dfb7e91