print array types as Array<T> instead of T[]
Summary:
`?string[]` is confusing, is it `Array<?string>` or `?Array<string>`? (it's the latter). we recommend the parametric style, even though it is more verbose, because it is easier to reason about.

the type printer was using the bracket notation, and actually gets the parens wrong... `Array<?string>` was printed as `?string[]`. this diff changes to the parametric style and fixes that issue as well.

Reviewed By: gabelevi

Differential Revision: D7456219

fbshipit-source-id: d3844daf3e0c26c3a8da3f9940e02b1a848400b8