Fix no-color error printer on degenerate locations
Summary:
This fixes a crash in the no-color error printer which would happen when
printing a degenerate location.

A degenerate location has a column position that is outside the line length.
This can happen due to Loc.btwn_exclusive, which sets the start position of the
created loc to the end position of another loc.

For example, for JSX children, we create such a location between the opening
and closing element of the parent element. E.g.,

```
<Foo>
  <Bar />
  <Baz />
</Foo>
```

In the above, the children location will start at (1, 6), after the end of the
line, and end at (4, 0).

Before this change, the printer would crash trying to create the hyphen part of
the overline, because the calculated length was negative. After this change, we
will create an overline/underline pair that prints like this:

```
     v
<Foo>
  <Bar />
  <Baz />
</Foo>
^ [1]
```

Arguably, the location of the children should be changed to look like this
instead:

```
<Foo>
  v------
  <Bar />
  <Baz />
  ------^ [1]
</Foo>
```

And `Loc.btwn_exclusive` should just go away, which would avoid the most likely
cause of degenerate locations like this.

Reviewed By: gabelevi

Differential Revision: D7463596

fbshipit-source-id: d235cd1a38512bbf4eee44689140f1a120623fb5