Make `RPolyTest` related locations consistent across all bounds
Summary:
This changes the locations that populate the reason associated with `RPolyTest`.
This reason description is used when creating the bounds of type parameters via
`Flow_js.generate_tests`. The bounds are the `Empty` type and the declared
bound of the parameter if there is one, or `mixed` otherwise.

Before, the locations associated with these generated types were not consistent.
The `loc` and `def_loc` for the `Empty` test was that the location the type parameter,
but in some cases the `loc` and `def_loc` of the upper bound was the location of the
bound type itself.

In order for the normalizer to be able to match these types back to the original type
parameter these locations need to match. So this is what this commit attempts to
do. In particular, for the top bound, we use
- 'loc': the location of the type parameter (may be repositioned later)
- 'def_loc': the location of the type parameter, and
- 'annot_loc_opt': the location of the bound (if present)

Reviewed By: samwgoldman

Differential Revision: D7385357

fbshipit-source-id: a7854779bf9308ce255fc4776b652983f4d239ff