[Normalizer] Delete the `Ty.This` constructor (will appear as a type parameter)
Summary:
The `Ty.This` type is superfluous in the normalized language as it in essence represents
an implicit type parameter. In this commit we remove this constructor from `Ty.t`. This way
we can treat the `this` type as a type parameter which more closely resembles the way it is
handled throughout type checking.

Reviewed By: samwgoldman

Differential Revision: D7397132

fbshipit-source-id: fe85350e98a40b22c36cd55d0b452e72f9e292dd