[Normalizer] Normalized type are serialized to AST nodes
Summary:
This converts normalized `Ty.t` types to (almost) `Ast.Type.t` nodes,
so they can later be printed reusing existing AST printer infrastructure.

This is a partial function. We restrict support for parts of `Ty.t` that are serializeable
to `Ast.Type.t`. So cases like `TypeAlias` or recursive types are excluded for now.

The immediate use case will be the "suggest" command.

Reviewed By: samwgoldman

Differential Revision: D6739677

fbshipit-source-id: 6f23fc7f76ee0bcf80e0a13bb65bab451a8f03e5