Add synchronous APIs to monitor logger
Summary:
When the monitor logs using `FlowServerMonitorLogger`, each log
command asynchronously writes the message to the log and or stderr. It
is good that it is asynchronous, so that the monitor doesn't go
unresponsive while it logs.

However, when we're debugging it's really useful to have synchronous
logging. If Lwt is acting strangely, I want to be able to using logging
to figure out what happens when.

Some lwt APIs use `_s` to indiciate synchronous APIs. So I've done that
here!

Reviewed By: samwgoldman

Differential Revision: D7464042

fbshipit-source-id: 39419da1aebf68f86e9ec5cba21a8d277186e1e8