Don't instantiate poly component class for createElement
Summary:
Before this diff, we used the resolved component type. This was a problem when
the component class was itself polymorphic. Resolving the PolyT would create an
implicitly-instantiated InstanceT, which contains the tvars from implicit
instantiation. These tvars would be reachable through the `type` field of the
element.

This diff still resolves the component class, but traps before instantiating
the typeapp, if there is one.

Reviewed By: avikchaudhuri

Differential Revision: D6973260

fbshipit-source-id: ad2280084d7a9fe2da36dda639e6885c761bbf3b