Change AnnotT back to wrapping a Type.t
Summary:
In a follow-up diff, I make a change where all `typeof` types are wrapped in
`AnnotT`. This wrapper is useful because downstream code can better detect
whether a given type is an annotation or not.

This would not be important, except that `typeof` is problematic. The
implementation waits for the `typeof`-ed type to resolve to a concrete type,
but type type might recursively contain tvars.

These tvars confound later analysis, like concretization for speculation, but
also assert_ground, which walks types and raises missing annotation errors on
tvars.

Reviewed By: avikchaudhuri

Differential Revision: D7489344

fbshipit-source-id: 483aebc9593bcdd40ed8f9441688809737f18d1b