Use polarity sensitive type visitor to implement assert_ground
Summary:
The current visitor logic in assert_ground is a bit dense. In practice, we
should be asking for annotations on any tvar that appears in an input position,
which the current visitor approximates with an `infer` flag. However, the flag
doesn't properly respect polarities (does not flip back to positive on nested
function parameters, for example).

This diff uses the polarity-sensitive type visitor to walk exports instead, and
adds errors in the expected places.

There are several exceptions, which can be exported without annotations:
* object and array literals
* initialized class properties (instance & static)
* "munged" class properties
* shadow properties of unsealed objects
* existential tvars
* `this`

Reviewed By: panagosg7

Differential Revision: D6023275

fbshipit-source-id: 94631eb3702e00c6c90d7ff0231a23a43b37bb34