Skip over properties that have no location information attached
Summary:
As far as I am aware, this causes no observable change in behavior.

```
type Foo = {...{bar: string}};
const x: Foo = {bar: ''};
x.bar;
```

Given this code, when the `member` hook fires on `x.bar`, the receiver type that
ends up getting returned is an `EvalT`, which we already ignore in find-refs.
Due to this same issue, `get-def` does not work on the `bar` in `x.bar`. This
needs to be addressed but that's a battle for a different day.

So, even though this does not currently change behavior, it is necessary for
D7386313, which adds object literal properties to find-refs results. In that
diff we use the `objt_to_objt` hook and examine objects into which object
literals flow. In this example, when the object literal flows into `Foo`, we get
an actual `ObjT`, as opposed to the `EvalT` we get from the member hook. When we
examine that type we go through the code changed in this diff, but the spreaded
properties from that type do not have locations. This means that when we get the
error that this diff removes, and the whole find-refs request blows up.

Note that this error gets triggered even when the type turns out to be unrelated
to the original request. We use this codepath to examine all candidates and
determine which ones are relevant.

Reviewed By: panagosg7

Differential Revision: D7459635

fbshipit-source-id: d6ac05d38f4cf4bd5c465a0ceb8856fcdd3be9df