shortcut for negative number literals
Summary:
before, the unary minus operator was handled such that `-x` corresponds to `type of x ~> UnaryMinusT (tout)`, regardless of what `x` is. `__flow` then recognizes the `NumT (Literal) ~> UnaryMinusT` case and multiplies the literal value by -1.

this means that the type of the AST node `-1` is an unresolved tvar rather than a `NumT (Literal -1)`. in a few places, we pattern match on these syntactically generated types directly, rather than via constraints. one such place is when `Object.freeze` generates the resulting object type; if a prop's value type is a `NumT`, we generate a `SingletonNumT`, but if it's an OpenT then we don't. if we don't, then the `NumT` gets summarized to `number` rather than the singleton value.

by example: `let x = Object.freeze({ foo: -1, bar: 2 }); type T = $Values<typeof x>`
the `-1` generates a `NumT ~> UnaryMinusT (tout)`, and the `$Values` generates a `obj ~> GetValuesT`. the latter special cases literal values in frozen objects, but in this case the `tout` is not a literal, even though it will always resolve to exactly one. so we get `type T = number | 2`.

anyway, this diff special cases the unary minus on a number literal to directly create a NumT instead of a tvar, so that the above special case generates a SingletonNumT instead of generalizing.

in this example, now `type T = -1 | 2`

Reviewed By: samwgoldman

Differential Revision: D7506793

fbshipit-source-id: fc995d9c1468ed94bc9744ccb2780e0af7311092