AST record label changes
Summary:
The AST data type has record labels in camelCase, which clashes with the rest
of the codebase, which is written in a snake_case style.

This diff also clarifies the distinction between type parameters and type
arguments, which both used the `typeParameters` record label.

Finally, this diff uses abbreviations which are used pervasively elsewhere,
both in our own code and in the industry at large.

Summary of changes:
* typeAnnotation -> annot
* returnType -> return
* typeParameters -> tparams (when actually params)
* typeParameters -> targs (when actually args)
* superClass -> super
* superTypeParameters -> super_targs (actually args here)

There are a few camelCase stragglers, but they are much more niche, like JSX
openingElement. I'm happy to do these as well, but kind of lost steam on this
diff.

Reviewed By: mroch

Differential Revision: D7524906

fbshipit-source-id: bf51ced13f1eabc1167842243ba4f990507178c6