Improve Promise typing
Summary:
Flow comes with a definition for Promise built in, but www overrides this
definition to add a non-standard `done` method. It _also_ changes the
types of of `then` and `catch`, to allow passing `null` as the `onFulfill` and
`onReject` callbacks.

On the Flow side, I am trying to add a `PromiseLike` interface, but it's only
possible if the Flow definition and the www definition agree on what `then`
means. My goal is to come up with better types in www, then upstream the
definition into Flow.

Unfortunately, the definition in www has a big issue, which causes Flow to miss
errors. The implementation in this diff resolves those issues and ~60 errors
are suppressed in this diff.

The crux of the issue can be reduced to a simpler overload:

```
declare class P<T> {
  then<U>(cb?: T => U): P<U>;
}
declare var p: P<string>;
p.then().then(x => (x: empty));
```

When `cb` is not provided, the type parameter `U` is unconstrained. At runtime,
the Promise will actually return `P<T>` when `cb` is not a function. The change
here in terms of the example above is this:

```
declare class P<T> {
  then(cb: null|void): P<T>;
  then<U>(cb: T => U): P<U>;
}
declare var p: P<string>;
p.then().then(x => (x: empty));
```

Flow's overload resolution will pick the first definition and return
`P<string>` from the `.then()` call, so we find the `string ~> empty` error.

Note that Flow's overload resolution will always pick the first overload when
the `cb` type is `any`. This can actually cause spurious errors. If the
argument is `any`, we will pick the first override, which might be unexpected.

Reviewed By: avikchaudhuri

Differential Revision: D7518089

fbshipit-source-id: 78a3237d411879fcfd79d321d8594ed2fd4fc4fb