Add createContext, createRef to default export object as well
Summary:
We do a thing where we export var all the React functions, but we also export
default an object with them all. This makes `import React from 'react'` as well
as `import * as React from 'react'` work.

We added createContext and createRef recently, but neglected to add them to the
default export object. This diff makes the new APIs available in the same way
the old APIs have been.

Reviewed By: avikchaudhuri

Differential Revision: D7526419

fbshipit-source-id: b0f55f4118d96c184dbf2977f60b8ace6e0770f9