[hack] Thread environment through interrupt_handler in call_with_interrupt
Summary:
The function that can run during `MultiWorker` invocation, `interrupt_handler`, is rather limited now: it can look at provided FDs and return one bit of information: whether to cancel. This is not going to be enough, and I augment it here with an environment that is threaded through all invocations (allowing to accumulate state) and returned at the end.

Besides that:

* replacing a bool with a named variant (`Cancel | Continue`)
* packaging three interrupt-related args into a single struct

Reviewed By: alexchow

Differential Revision: D7550946

fbshipit-source-id: e6bd10e81e6d4939c871eae028970c89d526093e