[type tables] tracking type parameters in scope
Summary:
**Problem:**  During normalization, types used as polymorphic bounds
(through `Flow_js.generate_tests`) need to be matched with the type parameters
that they instantiated in the first place. This link was lost since at no point were
these bounds replaced by their abstract counterpart. The only information that
was kept around was the name of the type parameter as argument of the `RPolyTest`
reason description. However, this information is not enough to reliably reconstruct
the type parameter after inference is done for two reasons:
1. Just the name of a parameter is not enough in the face of nested parameters with
the same name.
2. A type with `RPolyTest` reason might escape the scope where the type
parameter is well-defined (e.g. in `function f<T>(x: T) { return x; }`,
the bounds for `T` will escape the boundaries of `f`).

**Solution:** To fix this, when normalizing a type, we need to also provide the normalizer
with the set of type parameters (with location of definition) that are in scope at the location
where the type occurred.

This commit is the first step towards this goal: it sets up a structure in `scope.ml`,
called `tparam_entries` that contains at every time, the type parameters that are in
scope. This structure is kept up-to-date by adding the necessary type
parameters in `func_sig.ml` and `class_sig.ml` as constraint generation enters and
exits polymorphic environments.

Reviewed By: samwgoldman

Differential Revision: D7401236

fbshipit-source-id: f902ab6db5d9d974ff3653ec73a95964babbcf8b