[normalizer] Fix type parameter scopes (when handling RPolyTest)
Summary:
This commit targets the normalization of type parameter bounds. Until now, we assumed that all
these type parameters were defined in the scope where the type introduced. This did not
work well with type parameters that were recovered from their bound instantiations
through the `RPolyTest` reason, as these types may easily escape the realm where the
type parameters are defined.

To fix this behavior, we restrict well-formed types to those whose free type parameters
belong to a set of type params that accompanies the type (and originates from the context
in which the type was introduced).

To implement this:
* [`type_table.ml`] we introduce the notion of a **scheme** which is a type (`Type.t`) accompanied
by an environment of type parameters. This environment is in the form of a list of `(name, Loc.t)`
pairs. The `Loc.t` portion is the location where the type parameter is defined. The same name may
appear multiple times (which is the case of type parameter shadowing. A `(name, Loc.t)` pair, however,
may only appear once.
* [`ty_normalizer.ml - RPolyTest`] Crucially a type-bound with `RPolyTest` only restores the
respective `Ty.Bound` if its `(name, def_loc)` pair appears in the attendant environment. Otherwise,
the expanded bounds are preserved instead of being folded to the type parameter. This check is
performed in `Env.lookup_tparam`.

* [`ty_normalizer.ml - BoundT`] The normalizer also performs this check when handling `BoundT`
types. For now this check should never fail, but there is a case where this check is useful that will be
handled in a later commit.*

* Clients need to provide a `scheme`, instead of just a `Type.t` whenever available. The
old behavior for `from_type` assumes an empty tparam environment.

[*]  Briefly:
```
function f<T>(x: T) {
  function g<T>(y: T, z) {
    g(y, x);
  }
}
```
Asking Flow to infer the type for `z` would return `T`, but this is `f`'s `T` and not `g`'s.

Reviewed By: samwgoldman

Differential Revision: D7401316

fbshipit-source-id: e607d1d56457b05a9e671d4a5a56d4966f7fa4a5