If a newtest fails, write out its log
Summary:
I wrote this diff for the sake of automated CI servers, which typically record the stdout+stderr of failed tests, but don't give you access to the machine/image on which the test ran.

With this change, the stdout log of the failed test will include its `test.log` file.

Reviewed By: gabelevi

Differential Revision: D7314902

fbshipit-source-id: d2cabaaebc3866e1f719757bdcb0e56c720c0dec