lint errors during speculation
Summary:
Lint errors should not be considered speculation failures. But it is still important to preserve
them and report them on successful speculations.

This is exactly what happens with other deferred "actions": flow and unify constraints on unresolved
tvars are logged and then replayed if necessary. This diff reuses that infra for addressing the
above problem with lint errors.

Reviewed By: samwgoldman

Differential Revision: D7573985

fbshipit-source-id: 2a287d95307d998f9f79582e2d98963dd260b9ac