fix local build of flow-parser-bin when dist/ doesn't exist
Summary:
this wasn't working when the `dist/` directory doesn't exist. the implicit rules were handling the cases where the zip files were built locally or downloaded from github, but not the case where the resulting libflowparser.a already exists and doesn't need the zip files, as in the npm package.

so this adds a `.premade` file, which we write when we pack the npm module. when that file exists, libflowparser.a is always considered up to date.

Reviewed By: avikchaudhuri, panagosg7

Differential Revision: D7528895

fbshipit-source-id: 0a8e547528297f72d74c3a695e5917aff5020d85