Rework flow IDE tests in preparation for LSP
Summary:
(This diff has a lot of changes. I've left them as one diff because it's hard to make changes to a stack of diffs. If the main direction looks good, I can split it up into smaller diffs for easier review)

(A) Rename `ideNewMessagesWithTimeout -> waitAndVerifyAllIDEMessageContentSinceStartOfStep` and `ideNoNewMessagesAfterSleep -> waitAndVerifyNoIDEMessagesSinceStartOfStep`. Why? (1) These combine an action waiting and an assertion verifying, and to support LSP's richer interactions, I'm going to have to add more actions-on-their-own and assertions-on-their-own, so I wanted to switch to a uniform naming scheme. (2) The fact that these only look at messages since the start of the step is so subtle that it needs to be called out in the name. (3) The fact that `ideNewMessagesWithTimeout` only works to assess the entirety of messages in the step and can't be composed with subsequent calls is so subtle that it needs to be called out in the name. I also renamed `env.getIDEMessages` to `env.getIDEMessagesSinceStartOfStep`, correspondingly.

(B) Rename `ideRequest` to `ideRequestAndWaitForResponse`. That's after all what it does. Previously it wasn't clear that it waited. So, in ide/test.js, there was actually one test which called this and then AFTERWARDS had a wait of 10000ms until the response arrived. Obviously that wait should have been 0ms since it by definition already has arrived. Also previously it wasn't clear that this method has an unbounded timeout. For LSP tests moving forwards, I'm going to use a new method `ideRequest` which merely sends the request, and then I'm going to have a separate call to wait for the response with a timeout. And a new method `verifyAllIDEMessagesInStep` which goes only by method name, and lets you specify which methods you're happy to ignore.

(C) In RunTestSuite.js I changed the wording of the message "Testing flow ide is really tricky..." so that instead it leads with the actual error. In testRunner.js I changed the wording so that if a test fails, then in addition to reporting the suite+step that failed, it also directly tells you the path to the test.log that you can look in.

(D) There's a method `serverIsRunning` which takes a boolean for whether it expects the server to be running. I found the phrase `serverIsRunning(false)` to be a bit confusing, so I switched it to take `'running' | 'stopped'`. That feels more readable. I also renamed it to `verifyServerStatus('running' | 'stopped')`, and added a suite of related uniformally-named methods: `verifyIDEStatus, waitUntilServerStatus, waitUntilIDEStatus`. (the last two are actions, not assertions).

(E) Previously the `builder.ideNewMessagesWithTimeout` took an array of expected messages, but it ignored their content and only looked at the length. The behavior of this method is VERY subtle in terms of which actual received messages it looks at. I therefore added extensive comments, and changed it to take a number parameter for the expected count. Also, previously for purposes of keeping a count of actual messages that arrived, it used to count responses+notifications prior to being called but only notifications after. I changed it to count responses+notifications after it was called as well.

(F) I changed `ide.messages` to also include the method-name of responses. Also, there was something funky going on with the autocomplete response. Jsonrpc convention is that when a response comes back, it stores data in its "result" field. The legacy IDE service was actually storing an object there with a key named "result", and was storing its results inside that. So, it had `result.result.xyz`. I changed the ide test to expect that. Also, as part of normalizing how messages are treated, I made everything print out uniformly to the log `IDE >>notification` or `IDE <<notification` or `IDE >>request` or `IDE <<response`.

(G) I pulled out sanitization of responses (e.g. turning any `.flowVersion` field into a stubbed-out placeholder). Also, for LSP purposes, I needed more sanitization as well to accomodate relative paths.

(H) I changed `ideStart` to work with both legacy and LSP modes. My thought is that the existing test infrastructure was centrally designed for testing JSONRPC interactions with flow. Most of it worked fine with IDE and LSP. That's why I wanted to fit into the existing methods, rather than inventing an entirely parallel set of them. I also changed the sanity check in `runTestSuite.js`. Previously it required that you call waitAndVerifyNoIDEMessagesSinceStartOfStep or
waitAndVerifyAllIDEMessagesContentSinceStartOfStep on every step from the first `ideStart` to the last one. With LSP, merely invoking `ideStart` doesn't actually result in any messages, so I relaxed it to let you call `ideStart` on its own first.

(I) I added just the first lsp test to prove out all those changes. I left the rest in comments. I know it's not good to leave stuff in comments. But I'm going to be working on them immediately right this instant.

Reviewed By: gabelevi

Differential Revision: D7225780

fbshipit-source-id: 18a05088eb29b19f54c7a8069429bfe768b49548