Add LSP test for cold start
Summary:
This diff adds a second LSP integration test, for a cold start.

In adding this test I found I needed wanted a new TestStep API, one that would let me verify select content of messages.

I don't want to go to the full extent of waitAndVerifyAllIDEMessagesContentSinceStartOfStep, which requires a 100% match on the exact full content of every single message. It ends up firing even when there are irrelevant differences, e.g. different sequence numbers.

So instead I invented a mini-DSL. It uses simple string matches, e.g. if I ignore `window/progress{Connecting}` then it will ignore all progress messages where the word 'Connecting' appears anywhere in the content; if I expect `window/progress{null}` then it will match any progress message with that content.

(This test isn't yet complete: the server sends a request to the test harness, and the test harness is expected to deliver a response or respond to a cancellation request from the server. I'll leave that for a future diff.)

Reviewed By: gabelevi

Differential Revision: D7253720

fbshipit-source-id: 0fa89c415465fec3970b4aa735fe4e0d2feb7d3f