Add LSP testing of requests from server
Summary:
Previously we handled requests that newtests made to the `flow lsp`.

This diff adds handling of requests that `flow lsp` makes to newtests.

We need it for the case of "cold startup", where `flow lsp` sends a dialog request to newtests, and we want to verify that the dialog request was made, and also verify that the dialog request was cancelled.

Reviewed By: gabelevi

Differential Revision: D7265451

fbshipit-source-id: bfb4fff108a89809ef62e3b329f8f6b104423a95