Add LSP test for external restarting of flow server
Summary:
This test is for when the flow server stops for some reason (does LSP tell the user this fact?) and then the flow server gets restarted externally (does LSP dismiss its dialog and reconnect?)

It uses a new test API, `ideStartAndConnect`, which is a convenience rather than having to do several steps explicitly in each test. This API asserts that it has indeed connected correctly, and hence needs to return a TestStepSecondStage. I'd like to change `ideStart` to also return a TestStepSecondStage, so that a single API `ideStart` can do both (depending on what arguments you provide it). That would change a bunch of other tests, so I'll leave it for the future.

Reviewed By: gabelevi

Differential Revision: D7270216

fbshipit-source-id: 24306ad8a0d856c0cf3f6f1b911e3ea4a7a17f44