Add LSP test for internal restarting of flow server
Summary:
This diff adds a test for when flowLsp pops up a dialog box saying "Flow has stopped [Restart]" and the user clicks the button - it verifies that the server does indeed restart.

Along the way...

I added `ideResponse('mostRecent')` for responding to the most recent request from the server. This helps in cases where id of the server's request is non-deterministic, e.g. if it happens to take a longer time to connect.

I fixed a bug where test/builder.js was sending back the response `args` rather than `...args`.

I changed the reporting in the test log, so now the 'waitUntil' functions indicate in the log whether their condition was fulfilled or whether it timed-out before being fulfilled.

I remedied an omission in lspCommand where, if the user dismisses the dialog box saying "Flow is stopped", then it shouldn't pop up the dialog box upon subsequent failures to connect. (previously it would restore the dialog ever connection attempt).

I remedied an omission in lspCommand where, when it received a response from the client, it failed to dispatch that response to the appropriate local handler.

I remedied an omission in test/TestSetp.js where, if you `waitUntilServerStatus('stopped')`, then it should have set the flag `_allowServerToDie` so that when the server does indeed stop then the test will go on.

Reviewed By: gabelevi

Differential Revision: D7278560

fbshipit-source-id: f728f9ad55b76a8a12cca544fa48c18fee516e65