Add LSP test for modifying .flowconfig
Summary:
This diff adds a test for when there's a non-version change in .flowconfig (where we leave "flow lsp" still running and have it restart the server, because from the Nuclide/VSCode perspective it's really jolly undesirable for the LSP adapter to terminate).

And for when there's a version change in .flowconfig (where we have to make sure that "flow lsp" terminates because there's no other choice).

There were some other changes along the way...

MESSAGES OUTLIVE IDE. We want a test step where we verify that the IDE terminates, and we verify the sequence of messages that it had in this step prior to its termination. This means that `.messages` can no longer be a member of the `ide` structure (which becomes `null` upon IDE termination). I've moved it into a builder field called `ideMessages`.

MODIFY-FILE. We want this test to modify .flowconfig. I'd have preferred to just use `.addFile(source, '.flowconfig')`. But when .flowconfig is first created for the builder's directory, there's some custom logic to tweak the temp-dir setting in its .flowconfig. I had the choice to either modify `addFile` to be able to re-apply this tweak in future, or to add a new method `modifyFile` which bypasses that need. I figured that `modifyFile` was both cleaner and more generally useful.

LSPCOMMAND FETCH VERSION. There was a bug in lspCommand in its code for re-fetching the .flowconfig version to check for changes. This ended up not doing anything because .flowconfig is normally cached, and so `FlowConfig.get` is unable to check for changes. I added a new parameter to `FlowConfig.get` to indicate whether it's allowed to use the cache.

LSPCOMMEND CONNECTION STATUS. It used to report "telemetry/connectionStatus{false}" solely in `show_disconnected`. But in the case where there's a Server_fatal_exception and lspCommand determines (for reason of version channge) that it must terminate, then `show_disconnected` will never be called, and so we weren't getting that "connectionStatus{false}". I thought it would be better to have it in this case. So I made the Server_fatal_exception do the reporting itself.

NEW TEST API - DON'T MIND SERVER DEATH. In one of the tests, the flow server will shut down, but then be automatically restarted. I can't call `.waitUntilServerStatus('stopped')` because that runs the risk of being flakey, if the restart happens too quickly. But without calling that, the termination of the flow server results in a test exception saying that the flow server unexpectedly quit. So I added a new API `dontMindServerDeath()` to do the job of allowing the server to die in this step without complaint, but without actually doing any waiting.

Reviewed By: gabelevi

Differential Revision: D7285926

fbshipit-source-id: 55af0d403522bf36654dc25a8c843b721fc40a2f