Add LSP test for textDocument/definition
Summary:
This diff adds a test for textDocument/definition.

The current implementation in commandHandler to handle the definition request is at the moment just a placeholder, to be sure we're getting end-to-end testing. I had to tweak it to match LSP 0-based convention vs Flow mostly-1-based convention. Cleaning up those conversions will be left for a future diff.

Reviewed By: gabelevi

Differential Revision: D7288615

fbshipit-source-id: 1c1d3ffda82c59f2db7e003511ff3afd3a8b6e6f