LSP includes stack traces for Error cases
Summary:
When the client makes a request, it might end in OK, and ERROR
for case where failure is expected by the code, and EXCEPTION for cases
where there's a an expected logic bug in the code. Stack traces are useful
in tracking down anomolies in both Error (e.g. the code as written believes
it should be giving an empty answer but we dont believe it and need to track
down why), and in Exception. Traces already come for free with Exceptions
As for Errors,
we'll just need to manually add them in for each place that we think is useful.
This diff adds them in
only for the simplest of cases so far. I'll add them in where needed going
forwards.

Reviewed By: gabelevi

Differential Revision: D7336763

fbshipit-source-id: a0bf109ee62010452530fc6756fb69be9bba773a