Breaking change to context in --json-version 2
Summary:
I suppose I could bump the json-version to 3. Should I?

This affects the "context" property in every location. Currently it's
just the first line of the context, which makes printing multi-line
errors tricky. This updates it to include up to 5 lines of context. If
the context has the length 2N, it will include the first N and last N
lines. If the context has the length 2N+1, it will include the first N+1
and the last N lines.

Reviewed By: nmote

Differential Revision: D7585312

fbshipit-source-id: fefc26638636b432c714f516e0e77ab75e06b623