[hack] Daemon socket shouldnt reuse the same FD or leak
Summary:
Daemon socket shouldnt reuse the same FD or leak

Reusing same FD:
This is a long-standing problem. The old implementation was bad. Because we called "Unix.socketpair" only once (instead of Unix.pipe which is called twice), a process sees both ends of a socket. Which just isn't ergonomic. When doing "Unix.select" on a socket, you expect to be waiting for data coming from the other process, but in fact we could be seeing data that this own process wrote.

Don't leak FDs:
Leaking is bad. It also prevents EOF from appearing as you'd expect it to.

In a followup diff, WorkerController's Daemon.spawn will be changed to use `socket mode. Which makes it hit both of the above problems.

Reviewed By: gabelevi, gregory-nisbet

Differential Revision: D7495513

fbshipit-source-id: 18d86555d0eb2efa7745bdf270c86aeda637c686