[suggest] Rewrites command (prints a program after annotating it with inferred types)
Summary:
The suggest command was entirely broken, since it hadn't been carefully
ported to the new normalizer infrastructure.

This commit rewrites the command and adds a realistic test suite.

The command uses an AST visitor (`src/services/type_info/suggest.ml`) to discover
locations of interest that are not annotated. Currently these locations are function/method
parameters and return types. For each one of these locations it uses the inferred and
normalized type to fill-in the missing annotation.

To maximize success in discovering a location in the type tables, this command
queries both the table for coverage and type_info (see comment in `query_types.ml`).

The output program is produced using the pretty printer `src/parser_utils/output/printers/pretty_printer.ml`,
so there is no guarantee for textual equivalence of parts of the program that are not affected by the
annotation process.

Issuing:
```
$ flow suggest file.js
```
prints to stdout the original program annotated with inferred (and normalized)
types at places where annotations were missing originally.

When normalizing the inferred types there are several things that can go wrong:
* The inferred type could be `empty` (if for example a function is never called).
* The normalizer may reach an error state. (Remember the "normalize" function is not total.)
* The serializer (that transforms `Ty.t` to `Ast.Type.t`) might fail, since some of the
types that are inferred are not part of the AST type language (e.g. recursive types).

If any of these cases is encountered in normal mode, then the annotation will simply not be added.
If the `--fail-on-suggest-warnings` flag is passed then these cases will be printed as warnings
and no annotated program will be printed.

E.g. for:
```
function foo(x: number, y) {
  return x + y;
}
```
we get
```
flow suggest --fail-on-suggest-warnings tests/suggest/warn-empty-0.js

Warning ┈┈┈┈ tests/suggest/warn-empty-0.js:3:1

Inferred type is empty.

     1│ // flow
     2│
     3│ function foo(x: number, y) {
     4│   return x + y;
     5│ }
     6│

Warning ┈┈┈┈┈ tests/suggest/warn-empty-0.js:3:25

Inferred type is empty.

     1│ // flow
     2│
     3│ function foo(x: number, y) {
     4│   return x + y;
     5│ }
     6│

Found 2 warnings
```

In the default bahavior, Flow type errors do not affect our ability to produce an
annotated AST, so the command will work in such cases.

Passing the flag `--fail-on-tc-errors` will cause the command to fail if Flow encounters
errors or more than "max_warnings" warnings.

Parsing errors, however, are fatal and will cause the command to fail with an error,
since the AST will be severely distorted.

1. Syntax for non-annotated program nodes is not preserved, due to the use of the pretty printer.
2. The pretty printer does not preserve inline comments.

Reviewed By: gabelevi

Differential Revision: D7402709

fbshipit-source-id: fcb20fb9c2ab6467358729f95b22b9b402fc14a9